% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2func.R
\name{myboot2Lab9}
\alias{myboot2Lab9}
\title{My Bootstrap 2}
\usage{
myboot2Lab9(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iterations to be completed}

\item{x}{data to use}

\item{fun}{function to apply to data}

\item{alpha}{confidence interval adjuster}

\item{cx}{adjust character size}

\item{...}{alters the text on the histogram}
}
\value{
list and a plot
}
\description{
My Bootstrap 2
}
\examples{
\dontrun{myboot2(x = sample)}
}
