% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{My CLT Poisson}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{int, is a the sample size}

\item{iter}{int, is the number of times the random sample is run}

\item{lambda}{int is the lambda for the poisson function}

\item{...}{}
}
\value{
A plot with the arguments used of the CLT possion
}
\description{
My CLT Poisson
}
\examples{
{mycltp(n = 20, iter = 10000,)}
}
