% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycurve.R
\name{myncurve}
\alias{myncurve}
\title{My Normal Curve}
\usage{
myncurve(a, mu, sigma)
}
\arguments{
\item{a}{value to check P(Y <= a)}

\item{mu}{mean of the formula}

\item{sigma}{standard Deviation of the formula}
}
\value{
list with mu = mean, signma = standard deviation, area = area Y ~ N(x,y), P(Y <= a)
}
\description{
My Normal Curve
}
\examples{
{myncurve(3, 10, 5)}
{myncurve(1, 0.5, 0.1)}
}
