% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Number of Tickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats}

\item{gamma}{value (0.01-0.99) that the train is overbooked}

\item{p}{probability (0.00-0.99) of NOT overbooking}
}
\value{
a list: num tickets discrete, num tickets continuous, Num seats, gamma, probability confidence, and 2 plot in one
}
\description{
A function that find the ideal numbder of tickets for given values to sell that will be unlikely to be overbooked.
}
\examples{
ntickets(N = 400, gamma = 0.2, p = 0.95)
}
