% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnormFunc}
\alias{mymlnormFunc}
\title{My Maximum Likelyhood Normal}
\usage{
mymlnormFunc(x, mu, sig, ...)
}
\arguments{
\item{x}{is the sample vector}

\item{mu}{mean of the of the pop}

\item{sig}{signma of the pop}

\item{...}{extra stuff for the graph, used in contour}
}
\value{
list of data, and a plot
}
\description{
My Maximum Likelyhood Normal
}
\examples{
mymlnormFunc(x = c(10, 12, 13, 15, 12, 11, 10), mu = seq(10, 15, length = 1000), sig = seq(0.1, 4, length = 1000), lwd = 2, labcex = 1)
}
